# vw_gateway/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/user/login:
        post:
            tags:
                - Identity
            operationId: Identity_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/gateway.api.v1.user.id.LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.api.v1.user.id.LoginResp'
...
    /api/v1/user/register:
        post:
            tags:
                - Identity
            operationId: Identity_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/gateway.api.v1.user.id.RegisterRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.api.v1.user.id.RegisterResp'
    /api/v1/user/{userId}/logout:
        post:
            tags:
                - Identity
            operationId: Identity_Logout
            parameters:
                - name: userId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/gateway.api.v1.user.id.LogoutRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.api.v1.user.id.LogoutResp'
    /api/v1/userinfo/{userId}:
        get:
            tags:
                - Userinfo
            operationId: Userinfo_Userinfo
            parameters:
                - name: userId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.api.v1.user.userinfo.UserinfoResp'
    /api/v1/userinfo/{userId}/forgetPassword:
        post:
            tags:
                - Userinfo
            operationId: Userinfo_ForgetPassword
            parameters:
                - name: userId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/gateway.api.v1.user.userinfo.ForgetPasswordReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.api.v1.user.userinfo.ForgetPasswordResp'
    /api/v1/userinfo/{userId}/modifyEmail:
        post:
            tags:
                - Userinfo
            operationId: Userinfo_ModifyEmail
            parameters:
                - name: userId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/gateway.api.v1.user.userinfo.ModifyEmailReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.api.v1.user.userinfo.ModifyEmailResp'
    /api/v1/userinfo/{userId}/modifyPassword:
        post:
            tags:
                - Userinfo
            operationId: Userinfo_ModifyPassword
            parameters:
                - name: userId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/gateway.api.v1.user.userinfo.ModifyPasswordReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.api.v1.user.userinfo.ModifyPasswordResp'
    /api/v1/userinfo/{userId}/modifySignature:
        post:
            tags:
                - Userinfo
            operationId: Userinfo_ModifyUserSignature
            parameters:
                - name: userId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/gateway.api.v1.user.userinfo.ModifySignatureReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.api.v1.user.userinfo.ModifySignatureResp'
    /api/v1/userinfo/{userId}/modifyUsername:
        post:
            tags:
                - Userinfo
            operationId: Userinfo_ModifyUsername
            parameters:
                - name: userId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/gateway.api.v1.user.userinfo.ModifyUsernameReq'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.api.v1.user.userinfo.ModifyUsernameResp'
...
    /video/{videoId}/info:
        get:
            tags:
                - VideoInfo
            operationId: VideoInfo_GetVideoInfo
            parameters:
                - name: videoId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.api.v1.video.videoinfo.GetVideoInfoResp'
...
components:
    schemas:
...
        gateway.api.v1.user.id.LoginRequest:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
            description: 用户登录所需请求、响应
        gateway.api.v1.user.id.LoginResp:
            type: object
            properties:
                statusCode:
                    type: integer
                    description: app status code,The same below
                    format: int32
                msg:
                    type: string
                    description: app response message,The same below
                data:
                    $ref: '#/components/schemas/gateway.api.v1.user.id.LoginResp_Data'
        gateway.api.v1.user.id.LoginResp_Data:
            type: object
            properties:
                accessToken:
                    type: string
                refreshToken:
                    type: string
            description: when users login successfully, return two tokens to them
        gateway.api.v1.user.id.LogoutRequest:
            type: object
            properties:
                userId:
                    type: string
            description: 用户注销所需请求、响应
        gateway.api.v1.user.id.LogoutResp:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                msg:
                    type: string
        gateway.api.v1.user.id.RegisterRequest:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
                repeatPassword:
                    type: string
                gender:
                    type: integer
                    format: int32
                email:
                    type: string
                birthday:
                    type: string
                signature:
                    type: string
                inputCode:
                    type: string
            description: 用户注册所需请求、响应
        gateway.api.v1.user.id.RegisterResp:
            type: object
            properties:
                statusCode:
                    type: integer
                    format: int32
                msg:
                    type: string
                data:
                    $ref: '#/components/schemas/gateway.api.v1.user.id.RegisterResp_Data'
        gateway.api.v1.user.id.RegisterResp_Data:
            type: object
            properties:
                accessToken:
                    type: string
                refreshToken:
                    type: string
        gateway.api.v1.user.userinfo.CommonResp:
            type: object
            properties:
                status:
                    type: integer
                    format: int32
                message:
                    type: string
        gateway.api.v1.user.userinfo.ForgetPasswordReq:
            type: object
            properties:
                userId:
                    type: string
                email:
                    type: string
                newPassword:
                    type: string
                inputCode:
                    type: string
        gateway.api.v1.user.userinfo.ForgetPasswordResp:
            type: object
            properties:
                common:
                    $ref: '#/components/schemas/gateway.api.v1.user.userinfo.CommonResp'
        gateway.api.v1.user.userinfo.ModifyEmailReq:
            type: object
            properties:
                userId:
                    type: string
                email:
                    type: string
                inputCode:
                    type: string
        gateway.api.v1.user.userinfo.ModifyEmailResp:
            type: object
            properties:
                email:
                    type: string
                common:
                    $ref: '#/components/schemas/gateway.api.v1.user.userinfo.CommonResp'
        gateway.api.v1.user.userinfo.ModifyPasswordReq:
            type: object
            properties:
                userId:
                    type: string
                oldPassword:
                    type: string
                newPassword:
                    type: string
        gateway.api.v1.user.userinfo.ModifyPasswordResp:
            type: object
            properties:
                common:
                    $ref: '#/components/schemas/gateway.api.v1.user.userinfo.CommonResp'
        gateway.api.v1.user.userinfo.ModifySignatureReq:
            type: object
            properties:
                userId:
                    type: string
                signature:
                    type: string
        gateway.api.v1.user.userinfo.ModifySignatureResp:
            type: object
            properties:
                newSignature:
                    type: string
                common:
                    $ref: '#/components/schemas/gateway.api.v1.user.userinfo.CommonResp'
        gateway.api.v1.user.userinfo.ModifyUsernameReq:
            type: object
            properties:
                userId:
                    type: string
                newUsername:
                    type: string
        gateway.api.v1.user.userinfo.ModifyUsernameResp:
            type: object
            properties:
                newUsername:
                    type: string
                common:
                    $ref: '#/components/schemas/gateway.api.v1.user.userinfo.CommonResp'
        gateway.api.v1.user.userinfo.UserInfo:
            type: object
            properties:
                userName:
                    type: string
                email:
                    type: string
                signature:
                    type: string
                shells:
                    type: integer
                    format: int32
                cntFans:
                    type: integer
                    format: int32
                cntFollows:
                    type: integer
                    format: int32
                cntVideos:
                    type: integer
                    format: int32
                avatarPath:
                    type: string
                gender:
                    type: integer
                    format: int32
                birthday:
                    type: string
        gateway.api.v1.user.userinfo.UserinfoResp:
            type: object
            properties:
                userInfo:
                    $ref: '#/components/schemas/gateway.api.v1.user.userinfo.UserInfo'
                common:
                    $ref: '#/components/schemas/gateway.api.v1.user.userinfo.CommonResp'
...
        gateway.api.v1.video.videoinfo.GetVideoInfoResp:
            type: object
            properties:
                title:
                    type: string
...